% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_brain.R
\name{plot_brain}
\alias{plot_brain}
\title{Simple macaque brain plot}
\usage{
plot_brain(
  values = NULL,
  regions = NULL,
  base_image = NULL,
  color = NULL,
  color.family = c("viridis", "brewer"),
  color.theme = NULL,
  color.limits = NULL,
  legend = FALSE,
  labels = FALSE,
  legend.title = NULL
)
}
\arguments{
\item{values}{Vector of values to be plotted. Names are optional and should correspond with region names. If names are not provided, \code{regions} must be provided.}

\item{regions}{Vector of region names. Should be the same length as \code{values}. Required if \code{values} is an unnamed vector.}

\item{base_image}{Base image to be used. Currently, only the default \code{"chiou&decasien2022"} is supported.}

\item{color}{If \code{TRUE}, \code{values} are interpreted as color values to be plotted.}

\item{color.family}{Source of color palettes. Should be one of \code{c("brewer", "viridis")}.}

\item{color.theme}{Specific palette supported by one of the \code{color.family} sources.}

\item{color.limits}{Limits for the color range. It \code{NULL}, limits will be calculated from the data.}

\item{legend}{If \code{TRUE}, a legend will be included.}

\item{labels}{If \code{TRUE}, brain regions are labeled.}

\item{legend.title}{Custom title for the legend. Ignored if \code{legend} is not \code{TRUE}.}
}
\value{
A ggplot object.
}
\description{
This function plots data onto a simple macaque brain.
}
